split up tokenfactory file
